requirements.txt
# azure_func/requirements.txt
azure-functions==1.20.0
pandas
numpy
scipy
scikit-learn
joblib
